requirements.txt
Django==5.2.8
djangorestframework==3.16.1
psycopg[binary]==3.2.1
drf-spectacular==0.27.2
python-decouple==3.8
python-dotenv==1.0.0
